% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMeans.R
\name{AddMeans}
\alias{AddMeans}
\title{Adds the mean of the outcome variable for unit-time pairs where the variable whose coefficient is normalized is non-zero.}
\usage{
AddMeans(df, normalization_column, policyvar, outcomevar, custom_scalar = NULL)
}
\arguments{
\item{df}{Data frame containing the additional variables created for estimation.}

\item{normalization_column}{Variable whose coefficient will be normalized.}

\item{policyvar}{Policy variable.  # Another function uses "A string with the name of the policy variable used in \code{\link[=EventStudy]{EventStudy()}}." for this description. It might be good to use this more detailed description and be consistent across functions. - MZW}

\item{outcomevar}{Outcome variable.  # Same as above, a more detailed description would be more helpful. - MZW}

\item{custom_scalar}{Scalar to be displayed along y = 0 line if the user doesn't want the mean computed by the package.}
}
\description{
Adds the mean of the outcome variable for unit-time pairs where the variable whose coefficient is normalized is non-zero.
}
