% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddSmPath.R
\name{AddSmPath}
\alias{AddSmPath}
\title{Calculates smoothest path}
\usage{
AddSmPath(
  df,
  coefficients,
  inv_covar,
  conf_level = 0.95,
  maxorder = 10,
  maxiter_solver = 1e+06
)
}
\arguments{
\item{df}{Dataset with coefficients prepared for plotting (must include coefficients normalized in estimation).}

\item{coefficients}{Event-study coefficients (must include coefficients normalized in estimation).}

\item{inv_covar}{Inverse of covariance matrix of coefficients (must include row and column of zeros for normalized coefficients).}

\item{conf_level}{Confidence level to define critical value of Wald region. Should be a real number between 0 and 1, inclusive. Defaults to 0.95.}

\item{maxorder}{Sets a maximum polynomial order that will be used when calculating lowest possible polynomial order. Should be a whole number. Defaults to 10.}

\item{maxiter_solver}{Sets the maximum number of iterations when searching for the smoothest path with minimum squared term in highest order coefficient. Should be a positive whole number. Defaults to 1e6.}
}
\value{
df with smoothest path added as a new column
}
\description{
Function to add smoothest path to dataframe with coefficients
}
