% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddSuptBand.R
\name{AddSuptBand}
\alias{AddSuptBand}
\title{Title}
\usage{
AddSuptBand(estimates, num_sim, conf_level, seed, eventstudy_coefficients)
}
\arguments{
\item{seed}{}
}
\value{

}
\description{
Title
}
\examples{

}
