% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_EventStudyOLS_example}
\alias{df_EventStudyOLS_example}
\title{Sample dataset used in the Examples section in the EventStudyOLS documentation}
\format{
A data frame with 2000 rows and 13 variables (lead and lags created using post = 1, pre = 1, overidpre = 2, overidpost = 2):
\describe{
\item{id}{variable to identify units}
\item{t}{variable to identify time periods}
\item{z}{policy variable}
\item{y_base}{the outcome variable}
\item{x_r}{control variable}
\item{z_fd}{first difference of z}
\item{z_fd_lead1}{1 period lead of first difference of z}
\item{z_fd_lead2}{2 period lead of first difference of z}
\item{z_fd_lead3}{3 period lead of first difference of z}
\item{z_fd_lag1}{1 period lag of first difference of z}
\item{z_fd_lag2}{2 period lag of first difference of z}
\item{z_lag3}{3 period lag of z}
\item{z_lead3}{3 period lead of z}
}
}
\source{
raw dataset can be found in the .zip archive in \url{https://data.nber.org/data-appendix/w29170/}. Code for creating the example dataset can be found in \verb{/data-raw/df_EventStudyOLS_example.R}
}
\usage{
df_EventStudyOLS_example
}
\description{
Sample dataset used in the Examples section in the EventStudyOLS documentation
}
\keyword{datasets}
