% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventStudyFHS.R
\name{EventStudyFHS}
\alias{EventStudyFHS}
\title{Runs FHS with optional fixed effects and clustering}
\usage{
EventStudyFHS(
  prepared_model_formula,
  prepared_data,
  idvar,
  timevar,
  FE,
  TFE,
  cluster
)
}
\arguments{
\item{prepared_model_formula}{A formula object created in PrepareModelForumla that is passed to EventStudy.}

\item{prepared_data}{Data frame containing all of the parameters required for EventStudy() plus leads and
lags of the first differenced policy variable and leads and lags of the policy variable.}

\item{idvar}{Variable indicating units, should be a character.}

\item{timevar}{Variable indicating time periods, should be a character.}

\item{FE}{Specifies if unit fixed-effects should be included. Defaults to TRUE.}

\item{TFE}{Specifies if time fixed-effects should be included. Defaults to TRUE.}

\item{cluster}{Specifies whether to use clustered errors by units. If FALSE, will use unclustered
heteroskedasticity-robust standard errors. Defaults to TRUE.}
}
\value{
An estimated model
}
\description{
Runs FHS with optional fixed effects and clustering
}
\examples{
library(dplyr)
library(magrittr)

# Prepare data
data <- df_sample_dynamic \%>\% select(y_base, z, id, t, x_r, eta_m)
results <- EventStudy(estimator = "FHS", data = data, outcomevar = "y_base", policyvar = "z",
idvar = "id", timevar = "t", controls = "x_r", proxy = "eta_m", proxyIV = "z_fd_lead3",
FE = TRUE, TFE = TRUE, post = 3, overidpost = 1, pre = 0, overidpre = 3, normalize = -1,
cluster = TRUE)

# Prepare formula
model_formula <-  PrepareModelFormula(estimator = "FHS", outcomevar = "y_base",
str_policy_fd = c("z_fd", "z_fd_lead2", "z_fd_lead3", "z_fd_lag1", "z_fd_lag2", "z_fd_lag3"),
str_policy_lead = "z_lead3", str_policy_lag = "z_lag4", controls = "x_r", proxy = "eta_m",
proxyIV = "z_fd_lead3")

EventStudyFHS(prepared_model_formula = model_formula,
prepared_data = results[[2]]$data, idvar = "id", timevar = "t", FE = TRUE, TFE = TRUE,
cluster = TRUE)

}
