% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventStudyPlot.R
\name{EventStudyPlot}
\alias{EventStudyPlot}
\title{Title}
\usage{
EventStudyPlot(
  data,
  estimates,
  CI = 0.95,
  Supt = 0.95,
  seed = 1234,
  Preeventcoeffs = TRUE,
  Posteventcoeffs = TRUE,
  Nozeroline = FALSE,
  Smpath
)
}
\arguments{
\item{Smpath}{}
}
\value{

}
\description{
Title
}
