% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Sample dataset obtained from the replication archive for Freyaldenhoven et al. (2021)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2000 rows and 12 columns.
}
\source{
Dataset in .dta format can be found in the .zip archive in \url{https://data.nber.org/data-appendix/w29170/}
}
\usage{
example_data
}
\description{
Sample dataset obtained from the replication archive for Freyaldenhoven et al. (2021)
}
\keyword{datasets}
