% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFirstDifferences.R
\name{GetFirstDifferences}
\alias{GetFirstDifferences}
\title{Adds first difference of a variable as a new column}
\usage{
GetFirstDifferences(df, groupvar = NULL, timevar, diffvar)
}
\arguments{
\item{df}{Data frame that will be modified}

\item{groupvar}{Optional grouping variable, should be a character.}

\item{timevar}{Variable indicating time periods, should be a character.}

\item{diffvar}{Variable whose first difference will be taken, should be a character.}
}
\description{
Adds first difference of a variable as a new column
}
\seealso{
\link[data.table]{shift}
}
