% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareLags.R
\name{PrepareLags}
\alias{PrepareLags}
\title{Adds lags of a variable as new columns}
\usage{
PrepareLags(df, groupvar = NULL, timevar, leadvar, nlags)
}
\arguments{
\item{df}{Data frame that will be modified}

\item{groupvar}{Optional grouping variable.}

\item{timevar}{Variable indicating time periods.}

\item{leadvar}{Variable whose lags will be added.}

\item{nleads}{Number of lags to be added.}
}
\description{
Adds lags of a variable as new columns
}
\examples{
\dontrun{
#Compute lags of minimum wage in a panel data of states
PrepareLeads(df, groupvar = "state", timevar = "year", leadvar = "minwage", 5)
}

}
\seealso{
\link[data.table]{shift}
}
