% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareLeads.R
\name{PrepareLeads}
\alias{PrepareLeads}
\title{Adds leads of a variable as new columns}
\usage{
PrepareLeads(df, groupvar = NULL, timevar, leadvar, leads)
}
\arguments{
\item{df}{Data frame that will be modified.}

\item{groupvar}{Character indicating column of optional grouping variable.}

\item{timevar}{Character indicating column of time periods.}

\item{leadvar}{Character indicating column of variable whose leads will be added.}

\item{leads}{Numeric vector specifying the leads to be computed.}
}
\value{
The passed data frame augmented with columns called leadvar_leadN, where leadvar is the value specified and N is the selected lead(s).
}
\description{
\code{PrepareLeads} adds future values of a variable for a given number of periods based upon the time variable.
}
\examples{
PrepareLeads(
    df_sample_dynamic,
    groupvar = "id",
    timevar = "t",
    leadvar = "z",
    leads = 1:5
)

}
\seealso{
\link[data.table]{shift}
}
