% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreparePlottingData.R
\name{PreparePlottingData}
\alias{PreparePlottingData}
\title{Orders the eventstudy coefficients and generates the x-axis labels}
\usage{
PreparePlottingData(
  df_tidy_estimates,
  policyvar,
  post,
  overidpost,
  pre,
  overidpre,
  normalization_column,
  proxyIV = NULL
)
}
\arguments{
\item{df_tidy_estimates}{A data.frame created from applying estimatr::tidy()
to the estimation output from EventStudy.
At a minimum, it contains a column called "term" with the name for the coefficient and a
column called "estimate" that contains the corresponding estimate. Should be a data.frame.}

\item{policyvar}{Variable indicating policy variable z, should be a character.}

\item{post}{The number of periods in the past before
which the past values of the policy are not supposed
to affect the value of the outcome. Should be a whole number.
Corresponds to M in equation (2) of Freyaldenhoven et al. (forthcoming).}

\item{overidpost}{Optional number of event times after "post" to be
included in estimation. Defaults to 1. Should be a whole number.
Corresponds to L_M in equation (2) of Freyaldenhoven et al. (forthcoming).}

\item{pre}{Number of periods in the future after which the future values
of the policy are not supposed to affect the value of the outcome today.
Should be a whole number. Corresponds to G in equation (2) of
Freyaldenhoven et al. (forthcoming).}

\item{overidpre}{Optional number of event times earlier than -"pre" to be included in estimation.
Defaults to "post" + "pre". Should be a whole number.
Corresponds to L_G in equation (2) of Freyaldenhoven et al. (forthcoming).}

\item{normalization_column}{The name of the column containing the coefficient that will
be set to 0 in the eventstudy plot. Should be a character.}

\item{proxyIV}{Variables to be used as an instrument. Should be a character. if NULL,
defaults to the strongest lead of the policy variable based on the first stage.}
}
\value{
A data.frame that contains the x-axis labels, y-axis estimates,
and optional plot aesthetics to be used in creating the eventstudy plot
}
\description{
Orders the eventstudy coefficients and generates the x-axis labels
}
\examples{

tidy_eventstudy_estimates <- estimatr::tidy(EventStudy(estimator = "OLS",
data = df_sample_dynamic, outcomevar = "y_base",
policyvar = "z", idvar = "id", timevar = "t",
controls = "x_r", FE = TRUE, TFE = TRUE,
post = 3, pre = 2, overidpre = 4, overidpost = 5,
normalize = - 3, cluster = TRUE, default_override = FALSE)[[1]])

PreparePlottingData(
df_tidy_estimates = tidy_eventstudy_estimates,
policyvar = "z",
post = 3,
overidpost = 5,
pre = 2,
overidpre = 4,
normalization_column = "z_fd_lead3",
proxyIV = NULL)

# If you would like to use IV regression:
data <- df_sample_dynamic[, c("y_base", "z", "id", "t", "x_r", "eta_m")]

tidy_eventstudy_estimates <- estimatr::tidy(EventStudy(estimator = "FHS",
          data = data,
          outcomevar = "y_base", policyvar = "z", idvar = "id",
          timevar = "t", controls = "x_r", proxy = "eta_m", FE = TRUE, TFE = TRUE, post = 1,
          overidpost = 2, pre = 1, overidpre = 2,
          normalize = -1, cluster = TRUE, default_override = FALSE)[[1]])

PreparePlottingData(
df_tidy_estimates = tidy_eventstudy_estimates,
policyvar = "z",
post = 3,
overidpost = 5,
pre = 2,
overidpre = 4,
normalization_column = "z_fd_lead1",
proxyIV = "z_fd_lead3")

}
