% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestLinear.R
\name{TestLinear}
\alias{TestLinear}
\title{This function can be used to perform tests of linear hypotheses.}
\usage{
TestLinear(estimates, test, pretrends = NA, leveling_off = NA)
}
\arguments{
\item{estimates, }{A list containing estimation results and model information.
Should be an output ofEventStudyOLS or EventStudyIV}

\item{test, }{The hypothesis to be estimated. Accepts inputs that can be passed to hypothesis.matrix
argument in linearHypothesis() function from car package.}

\item{pretrends, }{String containing name of variable}

\item{leveling_off, }{String containing name of variable}
}
\description{
This function can be used to perform tests of linear hypotheses.
}
\examples{
TestLinear(Estimates, "z_fd_lag1 = z_fd", pretrends = "z_lead6", leveling_off = "z_lag8")

}
