% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddCIs.R
\name{AddCIs}
\alias{AddCIs}
\title{Adds columns to data frame containing confidence intervals around provided estimates.}
\usage{
AddCIs(df_estimates, policyvar, eventstudy_coefficients, conf_level = 0.95)
}
\arguments{
\item{df_estimates, }{A data frame with columns for term, estimate, and standard error.}

\item{policyvar, }{A string with the name of the policy variable used in EventStudy().}

\item{eventstudy_coefficients, }{A list specifying the names of the columns that were not normalized in EventStudy().}

\item{conf_level, }{Confidence level used for confidence interval expressed as a real number between 0 and 1, inclusively. Defaults to 0.95.}
}
\description{
Adds columns to data frame containing confidence intervals around provided estimates.
}
\examples{
estimates <- EventStudy(estimator = "OLS", data = df_sample_dynamic, outcomevar = "y_base",
                        policyvar = "z", idvar = "id", timevar = "t",
                        controls = "x_r", FE = TRUE, TFE = TRUE,
                        post = 3, pre = 2, overidpre = 4, overidpost = 5,
                        normalize = - 3, cluster = TRUE)

df_estimates_tidy <- estimatr::tidy(estimates[[1]])

df_estimates_tidy <- AddCIs(df_estimates_tidy,
                            policyvar = "z",
                            eventstudy_coefficients = estimates[[2]]$eventstudy_coefficients,
                            conf_level = 0.95)


}
