% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMeans.R
\name{AddMeans}
\alias{AddMeans}
\title{Adds the mean of the outcome variable for unit-time pairs where the variable whose coefficient is normalized is non-zero.}
\usage{
AddMeans(df, normalization_column, policyvar, outcomevar, custom_scalar = NULL)
}
\arguments{
\item{df}{Data frame containing the additional variables created for estimation.}

\item{normalization_column}{Variable whose coefficient will be normalized.}

\item{policyvar}{A string with the name of the policy variable used in \code{\link[=EventStudy]{EventStudy()}}.}

\item{outcomevar}{Variable indicating outcome variable y, should be a character.}

\item{custom_scalar}{Scalar to be displayed along y = 0 line if the user doesn't want the mean computed by the package.}
}
\description{
Adds the mean of the outcome variable for unit-time pairs where the variable whose coefficient is normalized is non-zero.
}
