% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_sample_dynamic}
\alias{df_sample_dynamic}
\title{Sample dataset obtained from the replication archive for Freyaldenhoven et al. (forthcoming)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2000 rows and 12 columns.
}
\usage{
df_sample_dynamic
}
\description{
Dataset in .dta format can be found in the .zip archive in \url{https://data.nber.org/data-appendix/w29170/}
}
\keyword{datasets}
