% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventStudy.R
\name{EventStudy}
\alias{EventStudy}
\title{Estimates Equation (2) in Freyaldenhoven et al. (forthcoming)}
\usage{
EventStudy(
  estimator,
  data,
  outcomevar,
  policyvar,
  idvar,
  timevar,
  controls = NULL,
  proxy = NULL,
  proxyIV = NULL,
  FE = TRUE,
  TFE = TRUE,
  post,
  overidpost = 1,
  pre,
  overidpre = post + pre,
  normalize = -1 * (pre + 1),
  cluster = TRUE
)
}
\arguments{
\item{estimator}{Accepts one of "OLS" or "FHS". If "FHS" is specified, implements IV.}

\item{data}{The data frame that contains the variables of interest.}

\item{outcomevar}{Variable indicating outcomme variable y, should be a character.}

\item{policyvar}{Variable indicating policy variable z, should be a character.}

\item{idvar}{Variable indicating units, should be a character.}

\item{timevar}{Variable indicating time periods, should be a character.}

\item{controls}{Optional vector of controls q, should be a character.}

\item{proxy}{Variable that is thought to be affected by the confoud but not by the policy.
Should be specified if and only if estimator is specified as "FHS". Should be a character.}

\item{proxyIV}{Variables to be used as an instrument. For the case of a single proxy,
defaults to the strongest lead of the policy variable based on the first stage.
Should be specified if and only if estimator is specified as "FHS".
Should be a character}

\item{FE}{Specifies if unit fixed-effects should be included. Defaults to TRUE.}

\item{TFE}{Specifies if time fixed-effects should be included. Defaults to TRUE.}

\item{post}{The number of periods in the past before which the past values of the policy
are not supposed to affect the value of the outcome. Should be a positive integer. Corresponds to M in equation (2) of
Freyaldenhoven et al. (forthcoming).}

\item{overidpost}{Optional number of event times after "post" to be included in estimation. Defaults to 1.
Should be a positive integer. Corresponds to L_M in equation (2) of Freyaldenhoven et al. (forthcoming).}

\item{pre}{Number of periods in the future after which the future values of the policy are
not supposed to affect the value of the outcome today. Should be a positive integer. Corresponds to G in equation (2) of
Freyaldenhoven et al. (forthcoming).}

\item{overidpre}{Optional number of event times earlier than -"pre" to be included in estimation. Defaults to "post" + "pre".
Should be a positive integer. Corresponds to L_G in equation (2) of Freyaldenhoven et al. (forthcoming).}

\item{normalize}{Specifies the event-time coefficient to be normalized. Defaults to - pre - 1.}

\item{cluster}{Specifies whether to use clustered errors by units. If FALSE, will use unclustered
heteroskedasticity-robust standard errors. Defaults to TRUE.}
}
\value{
A list that contains the estimation output and an object containing the arguments passed to the function
}
\description{
Estimates Equation (2) in Freyaldenhoven et al. (forthcoming)
}
\examples{
EventStudy(estimator = "OLS", data = df_sample_dynamic, outcomevar = "y_base",
policyvar = "z", idvar = "id", timevar = "t",
controls = "x_r", FE = TRUE, TFE = TRUE,
post = 3, pre = 2, overidpre = 4, overidpost = 5, normalize = - pre - 1, cluster = TRUE)
}
