% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventStudyOLS.R
\name{EventStudyOLS}
\alias{EventStudyOLS}
\title{Runs OLS with optional fixed effects and clustering}
\usage{
EventStudyOLS(
  prepared_model_formula,
  prepared_data,
  idvar,
  timevar,
  FE,
  TFE,
  cluster
)
}
\arguments{
\item{prepared_model_formula}{A formula object created in PrepareModelForumla that is passed to EventStudy.}

\item{prepared_data}{Data frame containing all of the parameters required for EventStudy() plus leads and
lags of the first differenced policy variable and leads and lags of the policy variable.}

\item{idvar}{Variable indicating units, should be a character.}

\item{timevar}{Variable indicating time periods, should be a character.}

\item{FE}{Specifies if unit fixed-effects should be included. Defaults to TRUE.}

\item{TFE}{Specifies if time fixed-effects should be included. Defaults to TRUE.}

\item{cluster}{Specifies whether to use clustered errors by units. If FALSE, will use unclustered
heteroskedasticity-robust standard errors. Defaults to TRUE.}
}
\value{
An estimated model
}
\description{
Runs OLS with optional fixed effects and clustering
}
\examples{
EventStudyOLS(prepared_model_formula = PrepareModelFormula("OLS", "y_base",
c("z_fd", "z_fd_lead2", "z_fd_lead3", "z_fd_lag1", "z_fd_lag2"),
"z_lead3", "z_lag3", "x_r"),
prepared_data = df_EventStudyOLS_example,
idvar = "id",
timevar = "t",
FE = TRUE, TFE = TRUE, cluster = TRUE)
}
