% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventStudyPlot.R
\name{EventStudyPlot}
\alias{EventStudyPlot}
\title{Creates an Event-Study Plot Following the Suggestions in Freyaldenhoven et al. (forthcoming)}
\usage{
EventStudyPlot(
  estimates,
  xtitle = "Event time",
  ytitle = "Coefficient",
  ybreaks,
  conf_level = 0.95,
  Supt = 0.95,
  num_sim = 1000,
  seed = 1234,
  Addmean = FALSE,
  Preeventcoeffs = TRUE,
  Posteventcoeffs = TRUE,
  Nozeroline = FALSE,
  Smpath = NULL
)
}
\arguments{
\item{estimates}{The output from calling EventStudy(). Should be a list of length 2.}

\item{xtitle}{The title for the x-axis. Should be a string. Defaults to "Event time".}

\item{ytitle}{The title for the y-axis. Should be a string. Defaults to "Coefficient".}

\item{ybreaks}{A vector containing the desired breaks for the y-axis. Should be a numeric vector that contains 0.}

\item{conf_level}{Confidence level used for confidence interval expressed as a real number between 0 and 1, inclusively. Defaults to 0.95.}

\item{Supt}{The confidence level used for obtaining the sup-t bands critical value. Should be a real number between
0 and 1, inclusive. Defaults to .95.}

\item{num_sim}{The number of simulations used in generating the sup-t bands. Should be a natural number.}

\item{seed}{The pseudorandom state used to make drawing "random" numbers reproducible. Should be a natural number.
Defaults to 1234.}

\item{Addmean}{Adds the mean of the dependent variable in the period used for normalization. Should be TRUE or FALSE. Defaults to FALSE.}

\item{Preeventcoeffs}{If TRUE, uses pre and overidpre from estimates to test for pre-trends. Should be TRUE or FALSE. Defaults to TRUE.}

\item{Posteventcoeffs}{If TRUE, uses post and overidpost from estimates to test for leveling-off. Should be TRUE or FALSE. Defaults to TRUE.}

\item{Nozeroline}{Whether or not to plot a dashed horizontal line at y = 0. Should be TRUE or FALSE. Defaults to FALSE, meaning the line is plotted.}

\item{Smpath}{PLACE HOLDER}
}
\value{
The Event-Study plot as a gpplot2 object
}
\description{
Creates an Event-Study Plot Following the Suggestions in Freyaldenhoven et al. (forthcoming)
}
\examples{
eventstudy_estimates <- EventStudy(estimator = "OLS", data = df_sample_dynamic, outcomevar = "y_base",
policyvar = "z", idvar = "id", timevar = "t",
controls = "x_r", FE = TRUE, TFE = TRUE,
post = 3, pre = 2, overidpre = 4, overidpost = 5, normalize = - 3, cluster = TRUE)

EventStudyPlot(estimates = eventstudy_estimates,
xtitle = "Event time",
ytitle = "Coefficient",
ybreaks = c(-1.5, -.5, 0, .5, 1.5),
conf_level = .95,
Supt = .95,
num_sim = 1000,
seed = 1234,
Addmean = FALSE,
Preeventcoeffs = TRUE,
Posteventcoeffs = TRUE,
Nozeroline = FALSE,
Smpath = NULL)
}
