% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventStudyPlot.R
\name{EventStudyPlot}
\alias{EventStudyPlot}
\title{Creates an Event-Study Plot following the suggestions in \href{https://www.nber.org/system/files/working_papers/w29170/w29170.pdf}{Freyaldenhoven et al. (2021)}.}
\usage{
EventStudyPlot(
  estimates,
  xtitle = "Event time",
  ytitle = "Coefficient",
  ybreaks = NULL,
  conf_level = 0.95,
  Supt = 0.95,
  num_sim = 1000,
  seed = 1234,
  Addmean = FALSE,
  Preeventcoeffs = TRUE,
  Posteventcoeffs = TRUE,
  Addzeroline = TRUE,
  Smpath = FALSE
)
}
\arguments{
\item{estimates}{The output from calling \code{\link[=EventStudy]{EventStudy()}}. Should be a list of length 2.}

\item{xtitle}{The title for the x-axis. Should be a string. Defaults to "Event time".}

\item{ytitle}{The title for the y-axis. Should be a string. Defaults to "Coefficient".}

\item{ybreaks}{A vector containing the desired breaks for the y-axis.
Defaults to NULL, which means the breaks are computed automatically.
If custom breaks are selected with the \code{Addmean} argument set to TRUE, then the breaks must include zero.}

\item{conf_level}{Confidence level used for confidence interval
expressed as a real number between 0 and 1, inclusive. Defaults to 0.95.}

\item{Supt}{The confidence level used for obtaining the sup-t bands critical value.
Should be a real number between 0 and 1, inclusive. Defaults to .95.}

\item{num_sim}{The number of simulations used in generating the sup-t bands.
Should be a natural number. Defaults to 1000.}

\item{seed}{The pseudorandom state used to make drawing "random" numbers reproducible.
Should be a natural number. Defaults to 1234.}

\item{Addmean}{Adds the mean of the dependent variable in the period used for normalization.
Should be TRUE or FALSE. Defaults to FALSE.  # Can we change the following variable names into snake case like conf_level for consistency? - MZW}

\item{Preeventcoeffs}{If TRUE, uses pre and overidpre from estimates to test for pre-trends.
Should be TRUE or FALSE. Defaults to TRUE.}

\item{Posteventcoeffs}{If TRUE, uses post and overidpost from estimates to test for leveling-off.
Should be TRUE or FALSE. Defaults to TRUE.}

\item{Addzeroline}{Whether or not to plot a dashed horizontal line at y = 0.
Should be TRUE or FALSE. Defaults to TRUE, meaning the line is plotted.}

\item{Smpath}{Plot smoothest path of confounder that rationalizes event study coefficients.
Should be TRUE or FALSE. Defaults to FALSE.}
}
\value{
The Event-Study plot as a gpplot2 object
}
\description{
Creates an Event-Study Plot following the suggestions in \href{https://www.nber.org/system/files/working_papers/w29170/w29170.pdf}{Freyaldenhoven et al. (2021)}.
}
\examples{

# OLS

eventstudy_estimates_ols <- EventStudy(
  estimator = "OLS",
  data = df_sample_dynamic,
  outcomevar = "y_base",
  policyvar = "z",
  idvar = "id",
  timevar = "t",
  controls = "x_r",
  FE = TRUE,
  TFE = TRUE,
  post = 3,
  pre = 2,
  overidpre = 4,
  overidpost = 5,
  normalize = - 3,
  cluster = TRUE,
  anticipation_effects_normalization = TRUE
)

EventStudyPlot(
  estimates = eventstudy_estimates_ols,
  xtitle = "Event time",
  ytitle = "Coefficient",
  ybreaks = c(-1.5, -.5, 0, .5, 1.5),
  conf_level = .95,
  Supt = .95,
  num_sim = 1000,
  seed = 1234,
  Addmean = FALSE,
  Preeventcoeffs = TRUE,
  Posteventcoeffs = TRUE,
  Addzeroline = TRUE,
  Smpath = FALSE
)

# This OLS example gives a warning message: Removed 1 rows containing missing values (geom_segment).

# IV

eventstudy_estimates_fhs <- EventStudy(
  estimator = "FHS",
  data = df_sample_dynamic,
  outcomevar = "y_base",
  policyvar = "z",
  idvar = "id",
  timevar = "t",
  controls = "x_r",
  FE = TRUE,
  TFE = TRUE,
  post = 3,
  pre = 0,
  overidpre = 3,
  overidpost = 1,
  normalize = - 1,
  cluster = TRUE,
  proxy = "eta_m",
  anticipation_effects_normalization = TRUE
)

EventStudyPlot(
  estimates = eventstudy_estimates_fhs,
  xtitle = "Event time",
  ytitle = "Coefficient",
  ybreaks = seq(-5, 10, 5),
  conf_level = .95,
  Supt = .95,
  num_sim = 1000,
  seed = 1234,
  Addmean = FALSE,
  Preeventcoeffs = TRUE,
  Posteventcoeffs = TRUE,
  Addzeroline = TRUE,
  Smpath = FALSE
)

# This IV example gives a warning message: "Removed 2 rows containing missing values (geom_segment)."
}
