% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareLags.R
\name{PrepareLags}
\alias{PrepareLags}
\title{Adds lags of a variable as new columns}
\usage{
PrepareLags(df, groupvar = NULL, timevar, lagvar, lags)
}
\arguments{
\item{df}{Data frame that will be modified.}

\item{groupvar}{Optional grouping variable, should be a character.}

\item{timevar}{Variable indicating time periods, should be a character.}

\item{lagvar}{Variable whose lags will be added, should be a character.}

\item{lags}{Numeric vector specifying the lags to be computed.}
}
\description{
Adds lags of a variable as new columns
}
\examples{
PrepareLags(df_sample_dynamic, groupvar = "id", timevar = "t", lagvar = "z", lags = 1:5)

}
\seealso{
\link[data.table]{shift}
}
