% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareLags.R
\name{PrepareLags}
\alias{PrepareLags}
\title{Adds lags of a variable as new columns}
\usage{
PrepareLags(df, groupvar = NULL, timevar, lagvar, lags)
}
\arguments{
\item{df}{Data frame that will be modified}

\item{groupvar}{Optional grouping variable.}

\item{timevar}{Variable indicating time periods.}

\item{lagvar}{Variable whose lags will be added.}

\item{lags}{Numeric vector specifying the lags to be computed.}
}
\description{
Adds lags of a variable as new columns
}
\examples{
\dontrun{
#Compute lags of minimum wage in a panel data of states
PrepareLags(df, groupvar = "state", timevar = "year", lagvar = "minwage", 1:5)
}
}
\seealso{
\link[data.table]{shift}
}
