% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareLags.R
\name{PrepareLags}
\alias{PrepareLags}
\title{Adds lags of a variable as new columns}
\usage{
PrepareLags(df, groupvar = NULL, timevar, lagvar, lags)
}
\arguments{
\item{df}{Data frame that will be modified.}

\item{groupvar}{Character indicating column of optional grouping variable.}

\item{timevar}{Character indicating column of time periods.}

\item{lagvar}{Character indicating column of variable whose lags will be added.}

\item{lags}{Numeric vector specifying the lags to be computed.}
}
\description{
Adds lags of a variable as new columns
}
\examples{
PrepareLags(
    df_sample_dynamic,
    groupvar = "id",
    timevar = "t",
    lagvar = "z",
    lags = 1:5
)

}
\seealso{
\link[data.table]{shift}
}
