% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareLeads.R
\name{PrepareLeads}
\alias{PrepareLeads}
\title{Adds leads of a variable as new columns}
\usage{
PrepareLeads(df, groupvar = NULL, timevar, leadvar, leads)
}
\arguments{
\item{df}{Data frame that will be modified.}

\item{groupvar}{Optional grouping variable, should be a character.}

\item{timevar}{Variable indicating time periods, should be a character.}

\item{leadvar}{Variable whose leads will be added, should be a character.}

\item{leads}{Numeric vector specifying the leads to be computed.}
}
\description{
Adds leads of a variable as new columns
}
\examples{
PrepareLeads(
    df_sample_dynamic,
    groupvar = "id",
    timevar = "t",
    leadvar = "z",
    leads = 1:5
)
# same spacing comment as in PrepareLags() - MZW

}
\seealso{
\link[data.table]{shift}
}
