% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareLeads.R
\name{PrepareLeads}
\alias{PrepareLeads}
\title{Adds leads of a variable as new columns}
\usage{
PrepareLeads(df, groupvar = NULL, timevar, leadvar, leads)
}
\arguments{
\item{df}{Data frame that will be modified}

\item{groupvar}{Optional grouping variable.}

\item{timevar}{Variable indicating time periods.}

\item{leadvar}{Variable whose leads will be added.}

\item{leads}{Numeric vector specifying the leads to be computed.}
}
\description{
Adds leads of a variable as new columns
}
\examples{
\dontrun{
#Compute leads of minimum wage in a panel data of states
PrepareLeads(df, groupvar = "state", timevar = "year", leadvar = "minwage", 5)
}
}
\seealso{
\link[data.table]{shift}
}
