% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareModelFormula.R
\name{PrepareModelFormula}
\alias{PrepareModelFormula}
\title{Title}
\usage{
PrepareModelFormula(
  estimator,
  data,
  outcomevar,
  policyvar,
  idvar,
  timevar,
  controls = NULL,
  proxy = NULL,
  proxyIV = NULL,
  FE = TRUE,
  TFE = TRUE,
  M,
  LM = 1,
  G,
  LG = 1,
  normalize = -1,
  cluster = TRUE
)
}
\arguments{
\item{cluster}{}
}
\value{

}
\description{
Title
}
