% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareModelFormula.R
\name{PrepareModelFormula}
\alias{PrepareModelFormula}
\title{Prepares a formula object for use in \code{\link[=EventStudyOLS]{EventStudyOLS()}} or \code{\link[=EventStudyFHS]{EventStudyFHS()}}}
\usage{
PrepareModelFormula(
  estimator,
  outcomevar,
  str_policy_fd,
  str_policy_lead,
  str_policy_lag,
  controls = NULL,
  proxy = NULL,
  proxyIV = NULL
)
}
\arguments{
\item{estimator}{Accepts one of "OLS" or "FHS". If "FHS" is specified, implements IV estimator in Freyaldenhoven et al. 2019.}

\item{outcomevar}{Variable indicating outcome variable y, should be a character.}

\item{str_policy_fd}{Vector indicating leads and lags of first differenced policy variable z with the (pre +1)th term omitted for normalization, should be a string.}

\item{str_policy_lead}{Variable indicating the (pre + overidpre)th lead of the policy variable z, should be a character.}

\item{str_policy_lag}{Variable indicating the (post + overidpost)th lag of the policy variable z, should be a character.}

\item{controls}{Optional vector of controls q, should be a character.}

\item{proxy}{Variable that is thought to be affected by the confound but not by the policy.
Should be specified if and only if estimator is specified as "FHS". Should be a character.}

\item{proxyIV}{Variables to be used as an instrument. Should be specified if and only if estimator is specified as "FHS". Should be a character.}
}
\value{
A formula object to be passed to EventStudy
}
\description{
Prepares a formula object for use in \code{\link[=EventStudyOLS]{EventStudyOLS()}} or \code{\link[=EventStudyFHS]{EventStudyFHS()}}
}
\examples{
PrepareModelFormula(estimator = "OLS", outcomevar = "y_base",
                    str_policy_fd = c("z_fd_lead1", "z_fd_lead2", "z_fd_lead4",
                                      "z_fd_lead5", "z_fd_lead6", "z_fd_lag1",
                                      "z_fd_lag2", "z_fd_lag3", "z_fd_lag4",
                                      "z_fd_lag5", "z_fd_lag6", "z_fd_lag7"),
                    str_policy_lead = "z_lead6",
                    str_policy_lag = "z_lag8",
                    controls = "x_r")

# If you would like to use IV regression:
PrepareModelFormula(estimator = "FHS",
                    outcomevar = "y_base",
                    str_policy_fd = c("z_fd", "z_fd_lead2", "z_fd_lead3",
                                      "z_fd_lag1", "z_fd_lag2"),
                    str_policy_lead = "z_lead3",
                    str_policy_lag = "z_lag3",
                    controls = "x_r",
                    proxy = "eta_m",
                    proxyIV = "z_fd_lead3")

# The outcomes of these examples have a line that says "<environment: 0x00######>" - what is this for? - MZW
}
