% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareModelFormula.R
\name{PrepareModelFormula}
\alias{PrepareModelFormula}
\title{Prepares a formula object for use in EventStudyOLS or EventStudyFHS}
\usage{
PrepareModelFormula(
  estimator,
  outcomevar,
  str_policy_fd,
  str_policy_lead,
  str_policy_lag,
  controls = NULL,
  proxy = NULL,
  proxyIV = NULL
)
}
\arguments{
\item{estimator}{Accepts one of "OLS" or "FHS". If "FHS" is specified, implements IV estimator in Freyaldenhoven et al. 2019.}

\item{outcomevar}{Variable indicating outcome variable y, should be a character.}

\item{str_policy_fd}{Vector indicating leads and lags of first differenced policy variable z with the (pre +1)th term omitted for normalization, should be a string.}

\item{str_policy_lead}{Variable indicating the (pre + overidpre)th lead of the policy variable z, should be a character.}

\item{str_policy_lag}{Variable indicating the (post + overidpost)th lag of the policy variable z, should be a character.}

\item{controls}{Optional vector of controls q, should be a character.}
}
\value{
A formula object to be passed to EventStudy
}
\description{
Prepares a formula object for use in EventStudyOLS or EventStudyFHS
}
\examples{
PrepareModelFormula(estimator = "OLS", outcomevar = "y_base",
str_policy_fd = c("z_fd_lead1", "z_fd_lead2", "z_fd_lead4",
"z_fd_lead5", "z_fd_lead6", "z_fd_lag1", "z_fd_lag2",
"z_fd_lag3", "z_fd_lag4", "z_fd_lag5", "z_fd_lag6", "z_fd_lag7"),
str_policy_lead = "z_lead6",
str_policy_lag = "z_lag8",
controls = "x_r")
}
