% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestLinear.R
\name{TestLinear}
\alias{TestLinear}
\title{This function can be used to perform tests of linear hypotheses.}
\usage{
TestLinear(estimates, test = NA, pretrends = T, leveling_off = T)
}
\arguments{
\item{estimates, }{A list of length 2 containing estimation results and model information.
Should be an output of EventStudyOLS or EventStudyIV.}

\item{test, }{The hypothesis to be estimated. Accepts inputs that can be passed to hypothesis.matrix
argument in linearHypothesis() function from car package.}

\item{pretrends, }{If TRUE, uses pre and overidpre from estimates to test for pre-trends. Defaults to TRUE.}

\item{leveling_off, }{If TRUE, uses post and overidpost from estimates to test for leveling-off. Defaults to TRUE.}
}
\description{
This function can be used to perform tests of linear hypotheses.
}
\examples{
TestLinear(Estimates, test = "z_fd_lag1 = z_fd", pretrends = T, leveling_off = T)

}
